requirements.txt
pip
ipython

python-dotenv

requests
requests-cache

tabula-py
tabulate

lxml
beautifulsoup4